Extract `WeatherHealthAlarmState` into dedicated file
